Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
pymysql==1.1.0
cryptography==41.0.8
requests==2.31.0
aiohttp==3.9.1
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
slowapi==0.1.9
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.2
tenacity==8.2.3
pandas==2.1.4
numpy==1.25.2
dateparser==1.2.0
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
sentence-transformers==2.2.2
huggingface-hub==0.16.4
scikit-learn==1.3.2
# Hybrid Analysis Dependencies (OPTIONAL - has NumPy conflicts)
# spacy==3.7.2  # Conflicts with scikit-learn - requires numpy>=2.0
# transformers==4.35.2
# torch==2.1.1

